Fixed initialization of streams in spc1.c
It looks like spc1.c was returning to the caller without all
the streams initialized.  The problem was first seen when the
value returned by spc1_next_op() was zeros without any errors.

Also increased the read ratio.  THIS IS TEMPORARY.  I will be adding
code to set it back to default, and allow the caller to set their
own read ratio.

Signed-off-by: Luis Pabón <lpabon@redhat.com>